requirements.txt
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.10.0
